refactor: simplify navigation components
