+++ src/components/Layout.tsx
@@ -1,10 +1,10 @@
 import Footer from './Footer'
-import Header from './Header'
+import Navigation from './navigation'
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
     <>
-      <Header />
+      <Navigation />
       <main
         className="flex flex-col mx-auto max-w-6xl justify-center py-10 px-4"
         role="main"
