--- src/components/Logo/index.ts
@@ -1,3 +1,0 @@
-import Logo from './Logo'
-
-export default Logo
