+++ src/components/navigation/index.tsx
@@ -1,0 +1,3 @@
+import Navigation from './navigation'
+
+export default Navigation
